Updated to have basic flight declaration capability
