+++ scd_operations/views.py
@@ -7,10 +7,16 @@ from auth_helper.utils import requires_scopes
 from rest_framework.response import Response
 from dataclasses import asdict, is_dataclass
 from datetime import timedelta
+from .scd_test_harness_helper import (
+    conflict_with_flight_test_injection_response,
+    planned_test_injection_response,
+    ready_to_fly_injection_response,
+    rejected_test_injection_response,
+    failed_test_injection_response,
+)
 from .scd_data_definitions import (
     SCDTestInjectionDataPayload,
     FlightAuthorizationDataPayload,
-    TestInjectionResult,
     SCDTestStatusResponse,
     CapabilitiesResponse,
     DeleteFlightResponse,
@@ -103,55 +109,9 @@ def SCDClearAreaRequest(request):
     # my_redis_helper = RedisHelper()
     # my_redis_helper.delete_all_opints()
 
-    outline_polygon = None
-    outline_circle = None
-    if "outline_polygon" in extent_raw["volume"].keys():
-        all_vertices = extent_raw["volume"]["outline_polygon"]["vertices"]
-        polygon_verticies = []
-        for vertex in all_vertices:
-            v = LatLngPoint(lat=vertex["lat"], lng=vertex["lng"])
-            polygon_verticies.append(v)
-        # polygon_verticies.pop() # remove the last vertex to prevent interseaction
-
-        outline_polygon = Polygon(vertices=polygon_verticies)
-
-    if "outline_circle" in extent_raw["volume"].keys():
-        circle_center = LatLngPoint(
-            lat=extent_raw["volume"]["outline_circle"]["center"]["lat"],
-            lng=extent_raw["volume"]["outline_circle"]["center"]["lng"],
-        )
-        circle_radius = Radius(
-            value=extent_raw["volume"]["outline_circle"]["radius"]["value"],
-            units=extent_raw["volume"]["outline_circle"]["radius"]["units"],
-        )
-        outline_circle = Circle(center=circle_center, radius=circle_radius)
-
-    altitude_lower = Altitude(
-        value=extent_raw["volume"]["altitude_lower"]["value"],
-        reference=extent_raw["volume"]["altitude_lower"]["reference"],
-        units=extent_raw["volume"]["altitude_lower"]["units"],
-    )
-    altitude_upper = Altitude(
-        value=extent_raw["volume"]["altitude_upper"]["value"],
-        reference=extent_raw["volume"]["altitude_upper"]["reference"],
-        units=extent_raw["volume"]["altitude_upper"]["units"],
-    )
-    volume3D = Volume3D(
-        outline_circle=outline_circle,
-        outline_polygon=outline_polygon,
-        altitude_lower=altitude_lower,
-        altitude_upper=altitude_upper,
-    )
-
-    time_start = Time(
-        format=extent_raw["time_start"]["format"],
-        value=extent_raw["time_start"]["value"],
-    )
-    time_end = Time(
-        format=extent_raw["time_end"]["format"], value=extent_raw["time_end"]["value"]
-    )
-
-    volume4D = Volume4D(volume=volume3D, time_start=time_start, time_end=time_end)
+    # Create a list of Volume4D objects
+    my_operational_intent_parser = dss_scd_helper.OperationalIntentReferenceHelper()
+    volume4D = my_operational_intent_parser.parse_volume_to_volume4D(volume=extent_raw)
     my_geo_json_converter = dss_scd_helper.VolumesConverter()
 
     my_geo_json_converter.convert_volumes_to_geojson(volumes=[volume4D])
@@ -231,33 +191,7 @@ def SCDAuthTest(request, operation_id):
         my_scd_dss_helper = dss_scd_helper.SCDOperations()
         my_geo_json_converter = dss_scd_helper.VolumesConverter()
         my_volumes_validator = dss_scd_helper.VolumesValidator()
-        
-        # Set the responses to be used
-        failed_test_injection_response = TestInjectionResult(
-            result="Failed",
-            notes="Processing of operational intent has failed",
-            operational_intent_id="",
-        )
-        rejected_test_injection_response = TestInjectionResult(
-            result="Rejected",
-            notes="An existing operational intent already exists and conflicts in space and time",
-            operational_intent_id="",
-        )
-        planned_test_injection_response = TestInjectionResult(
-            result="Planned",
-            notes="Successfully created operational intent in the DSS",
-            operational_intent_id="",
-        )
-        conflict_with_flight_test_injection_response = TestInjectionResult(
-            result="ConflictWithFlight",
-            notes="Processing of operational intent has failed, flight not deconflicted",
-            operational_intent_id="",
-        )
-        ready_to_fly_injection_response = TestInjectionResult(
-            result="ReadToFly",
-            notes="Processing of operational intent has failed, flight not deconflicted",
-            operational_intent_id="",
-        )
+
         # Get the test data
         scd_test_data = request.data
         # Prase the flight authorization data set
@@ -288,12 +222,14 @@ def SCDAuthTest(request, operation_id):
             )
         # Set up initial data
         now = arrow.now()
-        # Create a index and see what is already in Blender
-        my_rtree_helper = rtree_helper.OperationalIntentsIndexFactory(
-            index_name=INDEX_NAME
-        )
-        # Create a index of existing opints
-        my_rtree_helper.generate_operational_intents_index(pattern="flight_opint.*")
+
+        # # Create a index and see what is already in Blender
+        # my_rtree_helper = rtree_helper.OperationalIntentsIndexFactory(
+        #     index_name=INDEX_NAME
+        # )
+        # # Create a index of existing opints
+        # my_rtree_helper.generate_operational_intents_index(pattern="flight_opint.*")
+
         # Initial data for subscriptions
         one_minute_from_now = now.shift(minutes=1)
         one_minute_from_now_str = one_minute_from_now.isoformat()
@@ -316,17 +252,19 @@ def SCDAuthTest(request, operation_id):
                 },
                 status=status.HTTP_400_BAD_REQUEST,
             )
-        
+
         # Create a list of Volume4D objects
         all_volumes: List[Volume4D] = []
         for volume in operational_intent_volumes:
-            v4D = my_operational_intent_parser.parse_volume_to_volume4D(volume = volume)
+            v4D = my_operational_intent_parser.parse_volume_to_volume4D(volume=volume)
             all_volumes.append(v4D)
 
         # Create a list of Volume4D objects
         all_off_nominal_volumes: List[Volume4D] = []
         for off_nominal_volume in operational_intent_off_nominal_volumes:
-            v4D = my_operational_intent_parser.parse_volume_to_volume4D(volume = off_nominal_volume)
+            v4D = my_operational_intent_parser.parse_volume_to_volume4D(
+                volume=off_nominal_volume
+            )
             all_off_nominal_volumes.append(v4D)
 
         # Get THe state of the test data is coming in
@@ -402,7 +340,7 @@ def SCDAuthTest(request, operation_id):
 
         my_geo_json_converter.convert_volumes_to_geojson(volumes=all_volumes)
         view_rect_bounds = my_geo_json_converter.get_bounds()
-        # End check flight authorization
+
         auth_token = my_scd_dss_helper.get_auth_token()
         try:
             assert "error" not in auth_token
@@ -418,6 +356,9 @@ def SCDAuthTest(request, operation_id):
                 ),
                 status=status.HTTP_200_OK,
             )
+        print('+++++++++++++++++++++++++++++++++')
+        print(scd_test_data['operational_intent']['state'])
+        print('+++++++++++++++++++++++++++++++++')
         # Operational intents valid and now send to DSS
         op_int_submission = my_scd_dss_helper.create_and_submit_operational_intent_reference(
             state=test_injection_data.operational_intent.state,
